Issue 3902 - Bug: Cluster node does not get newly created node level secret

+++ agreementbot/secret_updater.go
@@ -134,85 +134,86 @@ func (sm *SecretUpdateManager) CheckForUpdates(secretProvider secrets.AgbotSecre
 
 // after each nodescan update the list of node secret managed by the agbot
 func (sm *SecretUpdateManager) UpdateNodePolicySecrets(org string, exchPolsMetadata map[string]exchange.ExchangeBusinessPolicy, secretProvider secrets.AgbotSecrets, db persistence.AgbotDatabase, agProtocol string) error {
+	for policyName, dpol := range exchPolsMetadata {
 
-       for policyName, dpol := range exchPolsMetadata {
+		// Get the list of managed secrets for this policy from the DB.
+		secretNames, err := db.GetManagedPolicySecretNames(org, exchange.GetId(policyName))
+		if err != nil {
+			glog.Errorf(smlogString(fmt.Sprintf("Error retrieving managed secret list for %s, error: %v", policyName, err)))
+			continue
+		}
 
-                // Get the list of managed secrets for this policy from the DB.
-                secretNames, err := db.GetManagedPolicySecretNames(org, exchange.GetId(policyName))
-                if err != nil {
-                        glog.Errorf(smlogString(fmt.Sprintf("Error retrieving managed secret list for %s, error: %v", policyName, err)))
-                        continue
-                }
-
-                // Keep track of which secrets are being referenced so that unused secrets can be removed at the end.
-                referencedSecrets := make(map[string]bool)
+		// Keep track of which secrets are being referenced so that unused secrets can be removed at the end.
+		referencedSecrets := make(map[string]bool)
 
 		// Iterate the list of secret bindings in the policy to get the secret manager secret names.
-                for _, sb := range dpol.SecretBinding {
-                        // Iterate each bound secret
-                        for _, bs := range sb.Secrets {
-                                // Extract the secret manager secret name
-                                _, secretFullName := bs.GetBinding()
-                                referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
+		for _, sb := range dpol.SecretBinding {
+			// Iterate each bound secret
+			for _, bs := range sb.Secrets {
+				// Extract the secret manager secret name
+				_, secretFullName := bs.GetBinding()
+				referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
 
 				if !sb.EnableNodeLevelSecrets {
-                                	continue
-                        	}
+					continue
+				}
 
-                                secretUser, _, secretName, err := compcheck.ParseVaultSecretName(secretFullName, nil)
-                                if err != nil {
-                                        glog.Errorf(smlogString(fmt.Sprintf("unable to parse secret name %s, error: %v", secretFullName, err)))
-                                        continue
-                                }
+				secretUser, _, secretName, err := compcheck.ParseVaultSecretName(secretFullName, nil)
+				if err != nil {
+					glog.Errorf(smlogString(fmt.Sprintf("unable to parse secret name %s, error: %v", secretFullName, err)))
+					continue
+				}
 
-                                pName := exchange.GetId(policyName)
+				pName := exchange.GetId(policyName)
 
-                                // Use now as the last update time for secrets that dont exist yet.
-                                secretLastUpdateTime := time.Now().Unix()
+				// Use now as the last update time for secrets that dont exist yet.
+				secretLastUpdateTime := time.Now().Unix()
 
 				agList, err := db.FindAgreements([]persistence.AFilter{persistence.PolAFilter(policyName), persistence.UnarchivedAFilter()}, agProtocol)
 				for _, ag := range agList {
-					secretNode := ag.DeviceId
+					secretNode := exchange.GetId(ag.DeviceId)
 
 					if secretUser != "" {
-						referencedSecrets[fmt.Sprintf("%s/user/%s/node/%s/%s", org, secretUser, secretNode, secretName)] = true
+						secretFullName = fmt.Sprintf("user/%s/node/%s/%s", secretUser, secretNode, secretName)
+						referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
 					} else {
-						referencedSecrets[fmt.Sprintf("%s/node/%s/%s", org, secretNode, secretName)] = true
+						secretFullName = fmt.Sprintf("node/%s/%s", secretNode, secretName)
+						referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
 					}
 
-                                	// Get the secret's metadata, if it exists
-                                	sm, err := secretProvider.GetSecretMetadata(org, secretUser, secretNode, secretName)
-                                	if err != nil {
-                                        	// The secret should be stored in the table even if it doesnt exist, so that if it is created later
-                                        	// changes to it will be recognized.
-                                        	glog.Warningf(smlogString(fmt.Sprintf("unable to retrieve metadata for %s %s, error: %v", org, secretFullName, err)))
-                                        	//continue
-                                	} else {
-                                        	secretLastUpdateTime = sm.UpdateTime
-                                	}
+					// Get the secret's metadata, if it exists
+					sm, err := secretProvider.GetSecretMetadata(org, secretUser, secretNode, secretName)
+					if err != nil {
+						// The secret should be stored in the table even if it doesnt exist, so that if it is created later
+						// changes to it will be recognized.
+						glog.Warningf(smlogString(fmt.Sprintf("unable to retrieve metadata for %s %s, error: %v", org, secretFullName, err)))
+						//continue
+					} else {
+						secretLastUpdateTime = sm.UpdateTime
+					}
 
-                                	glog.V(5).Infof(smlogString(fmt.Sprintf("storing managed secret %v %v from %v/%v", org, secretFullName, org, pName)))
+					glog.V(5).Infof(smlogString(fmt.Sprintf("storing managed secret %v %v from %v/%v", org, secretFullName, org, pName)))
 
-                                	// Only secrets that have never been referenced before are added to the DB. DB rows that already exist will not be updated.
-                                	err = db.AddManagedPolicySecret(org, secretFullName, org, pName, secretLastUpdateTime)
-                                	if err != nil {
-                                        	glog.Errorf(smlogString(fmt.Sprintf("unable to persist secret %v %v from %v/%v, error: %v", org, secretFullName, org, pName, err)))
-                                	}
+					// Only secrets that have never been referenced before are added to the DB. DB rows that already exist will not be updated.
+					err = db.AddManagedPolicySecret(org, secretFullName, org, pName, secretLastUpdateTime)
+					if err != nil {
+						glog.Errorf(smlogString(fmt.Sprintf("unable to persist secret %v %v from %v/%v, error: %v", org, secretFullName, org, pName, err)))
+					}
 				}
-                        }
-                }
+			}
+		}
 
-                // Look for unreferenced secrets and remove them.
-                for _, secretName := range secretNames {
-                        if _, ok := referencedSecrets[secretName]; !ok {
+		// Look for unreferenced secrets and remove them.
+		for _, secretName := range secretNames {
+			if _, ok := referencedSecrets[secretName]; !ok {
 
-                                glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, policyName)))
-                                err = db.DeletePolicySecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(policyName))
-                                if err != nil {
-                                        glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
-                                }
-                        }
-                }
+				glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, policyName)))
+				err = db.DeletePolicySecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(policyName))
+				if err != nil {
+					glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
+				}
+			}
+		}
 
 	}
 	return nil
@@ -221,89 +222,90 @@ func (sm *SecretUpdateManager) UpdateNodePolicySecrets(org string, exchPolsMetad
 // periodically called to update the list of node secret managed by the agbot
 func (sm *SecretUpdateManager) UpdateNodePatternSecrets(org string, exchPatsMetadata map[string]exchange.Pattern, secretProvider secrets.AgbotSecrets, db persistence.AgbotDatabase, agProtocol string) error {
 
-       for patternName, dpol := range exchPatsMetadata {
+	for patternName, dpol := range exchPatsMetadata {
 
-                // Get the list of managed secrets for this pattern from the DB.
-                secretNames, err := db.GetManagedPatternSecretNames(org, exchange.GetId(patternName))
-                if err != nil {
-                        glog.Errorf(smlogString(fmt.Sprintf("Error retrieving managed secret list for %s, error: %v", patternName, err)))
-                        continue
-                }
+		// Get the list of managed secrets for this pattern from the DB.
+		secretNames, err := db.GetManagedPatternSecretNames(org, exchange.GetId(patternName))
+		if err != nil {
+			glog.Errorf(smlogString(fmt.Sprintf("Error retrieving managed secret list for %s, error: %v", patternName, err)))
+			continue
+		}
 
-                // Keep track of which secrets are being referenced so that unused secrets can be removed at the end.
-                referencedSecrets := make(map[string]bool)
+		// Keep track of which secrets are being referenced so that unused secrets can be removed at the end.
+		referencedSecrets := make(map[string]bool)
 
 		// Iterate the list of secret bindings in the policy to get the secret manager secret names.
-                for _, sb := range dpol.SecretBinding {
-                        // Iterate each bound secret
-                        for _, bs := range sb.Secrets {
-                                // Extract the secret manager secret name
-                                _, secretFullName := bs.GetBinding()
-                                referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
+		for _, sb := range dpol.SecretBinding {
+			// Iterate each bound secret
+			for _, bs := range sb.Secrets {
+				// Extract the secret manager secret name
+				_, secretFullName := bs.GetBinding()
+				referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
 
 				if !sb.EnableNodeLevelSecrets {
-                                	continue
-                        	}
+					continue
+				}
 
-                                secretUser, _, secretName, err := compcheck.ParseVaultSecretName(secretFullName, nil)
-                                if err != nil {
-                                        glog.Errorf(smlogString(fmt.Sprintf("unable to parse secret name %s, error: %v", secretFullName, err)))
-                                        continue
-                                }
+				secretUser, _, secretName, err := compcheck.ParseVaultSecretName(secretFullName, nil)
+				if err != nil {
+					glog.Errorf(smlogString(fmt.Sprintf("unable to parse secret name %s, error: %v", secretFullName, err)))
+					continue
+				}
 
-                                pName := exchange.GetId(patternName)
+				pName := exchange.GetId(patternName)
 
-                                // Use now as the last update time for secrets that dont exist yet.
-                                secretLastUpdateTime := time.Now().Unix()
+				// Use now as the last update time for secrets that dont exist yet.
+				secretLastUpdateTime := time.Now().Unix()
 
 				agList, err := db.FindAgreements([]persistence.AFilter{persistence.PatAFilter(patternName), persistence.UnarchivedAFilter()}, agProtocol)
 				for _, ag := range agList {
-					secretNode := ag.DeviceId
+					secretNode := exchange.GetId(ag.DeviceId)
 
 					if secretUser != "" {
-						referencedSecrets[fmt.Sprintf("%s/user/%s/node/%s/%s", org, secretUser, secretNode, secretName)] = true
+						secretFullName = fmt.Sprintf("user/%s/node/%s/%s", secretUser, secretNode, secretName)
+						referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
 					} else {
-						referencedSecrets[fmt.Sprintf("%s/node/%s/%s", org, secretNode, secretName)] = true
+						secretFullName = fmt.Sprintf("node/%s/%s", secretNode, secretName)
+						referencedSecrets[fmt.Sprintf("%s/%s", org, secretFullName)] = true
 					}
 
-                                	// Get the secret's metadata, if it exists
-                                	sm, err := secretProvider.GetSecretMetadata(org, secretUser, secretNode, secretName)
-                                	if err != nil {
-                                        	// The secret should be stored in the table even if it doesnt exist, so that if it is created later
-                                        	// changes to it will be recognized.
-                                        	glog.Warningf(smlogString(fmt.Sprintf("unable to retrieve metadata for %s %s, error: %v", org, secretFullName, err)))
-                                        	//continue
-                                	} else {
-                                        	secretLastUpdateTime = sm.UpdateTime
-                                	}
+					// Get the secret's metadata, if it exists
+					sm, err := secretProvider.GetSecretMetadata(org, secretUser, secretNode, secretName)
+					if err != nil {
+						// The secret should be stored in the table even if it doesnt exist, so that if it is created later
+						// changes to it will be recognized.
+						glog.Warningf(smlogString(fmt.Sprintf("unable to retrieve metadata for %s %s, error: %v", org, secretFullName, err)))
+						//continue
+					} else {
+						secretLastUpdateTime = sm.UpdateTime
+					}
 
-                                	glog.V(5).Infof(smlogString(fmt.Sprintf("storing managed secret %v %v from %v/%v", org, secretFullName, org, pName)))
+					glog.V(5).Infof(smlogString(fmt.Sprintf("storing managed secret %v %v from %v/%v", org, secretFullName, org, pName)))
 
-                                	// Only secrets that have never been referenced before are added to the DB. DB rows that already exist will not be updated.
-                                	err = db.AddManagedPatternSecret(org, secretFullName, org, pName, secretLastUpdateTime)
-                                	if err != nil {
-                                        	glog.Errorf(smlogString(fmt.Sprintf("unable to persist secret %v %v from %v/%v, error: %v", org, secretFullName, org, pName, err)))
-                                	}
+					// Only secrets that have never been referenced before are added to the DB. DB rows that already exist will not be updated.
+					err = db.AddManagedPatternSecret(org, secretFullName, org, pName, secretLastUpdateTime)
+					if err != nil {
+						glog.Errorf(smlogString(fmt.Sprintf("unable to persist secret %v %v from %v/%v, error: %v", org, secretFullName, org, pName, err)))
+					}
 				}
-                        }
-                }
+			}
+		}
 
-                // Look for unreferenced secrets and remove them.
-                for _, secretName := range secretNames {
-                        if _, ok := referencedSecrets[secretName]; !ok {
+		// Look for unreferenced secrets and remove them.
+		for _, secretName := range secretNames {
+			if _, ok := referencedSecrets[secretName]; !ok {
 
-                                glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, patternName)))
-                                err = db.DeletePatternSecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(patternName))
-                                if err != nil {
-                                        glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
-                                }
-                        }
-                }
+				glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, patternName)))
+				err = db.DeletePatternSecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(patternName))
+				if err != nil {
+					glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
+				}
+			}
+		}
 
 	}
 	return nil
 }
-
 
 // When policies are added, changed or deleted, the list of managed secrets in the DB might need to be updated.
 func (sm *SecretUpdateManager) UpdatePolicies(org string, exchPolsMetadata map[string]exchange.ExchangeBusinessPolicy, secretProvider secrets.AgbotSecrets, db persistence.AgbotDatabase) error {
@@ -388,11 +390,12 @@ func (sm *SecretUpdateManager) UpdatePolicies(org string, exchPolsMetadata map[s
 		// Look for unreferenced secrets and remove them.
 		for _, secretName := range secretNames {
 			if _, ok := referencedSecrets[secretName]; !ok {
-
-				glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, policyName)))
-				err = db.DeletePolicySecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(policyName))
-				if err != nil {
-					glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
+				if _, secretNode, _, err := compcheck.ParseVaultSecretName(exchange.GetId(secretName), nil); secretNode != "" {
+					glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, policyName)))
+					err = db.DeletePolicySecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(policyName))
+					if err != nil {
+						glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
+					}
 				}
 			}
 		}
@@ -494,11 +497,12 @@ func (sm *SecretUpdateManager) UpdatePatterns(org string, exchPatternMetadata ma
 		// Look for unreferenced secrets and remove them.
 		for _, secretName := range secretNames {
 			if _, ok := referencedSecrets[secretName]; !ok {
-
-				glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, patName)))
-				err = db.DeletePatternSecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(patName))
-				if err != nil {
-					glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
+				if _, secretNode, _, err := compcheck.ParseVaultSecretName(exchange.GetId(secretName), nil); secretNode != "" {
+					glog.V(5).Infof(smlogString(fmt.Sprintf("deleting managed secret %s from %s because it is no longer used", secretName, patName)))
+					err = db.DeletePatternSecret(exchange.GetOrg(secretName), exchange.GetId(secretName), org, exchange.GetId(patName))
+					if err != nil {
+						glog.Errorf(smlogString(fmt.Sprintf("unable to delete %s from secrets DB, error: %v", secretName, err)))
+					}
 				}
 			}
 		}
